dom/workers/test/serviceworkers/test_imagecache.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1202085 - Test that images from different controllers don't cached together</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content">
<iframe></iframe>
</div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  var iframe;
  function runTest() {
    iframe = document.querySelector("iframe");
    iframe.src = "/tests/dom/workers/test/serviceworkers/fetch/imagecache/register.html";
    window.onmessage = function(e) {
      if (e.data.status == "ok") {
        ok(e.data.result, e.data.message);
      } else if (e.data.status == "registrationdone") {
        iframe.src = "/tests/dom/workers/test/serviceworkers/fetch/imagecache/index.html";
      } else if (e.data.status == "result") {
        is(e.data.url, "image-40px.png", "Correct url expected");
        is(e.data.width, 40, "Correct width expected");
        iframe.src = "/tests/dom/workers/test/serviceworkers/fetch/imagecache/unregister.html";
      } else if (e.data.status == "unregistrationdone") {
        iframe.src = "/tests/dom/workers/test/serviceworkers/fetch/imagecache/postmortem.html";
      } else if (e.data.status == "postmortem") {
        is(e.data.width, 20, "Correct width expected");
        window.onmessage = null;
        ok(true, "Test finished successfully");
        SimpleTest.finish();
      }
    };
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.interception.enabled", true],
      ["dom.serviceWorkers.enabled", true],
      ["dom.serviceWorkers.testing.enabled", true],
    ]}, runTest);
  };
</script>
</pre>
</body>
</html>
